Update main.py
Updated FileHandler.write() and FileHandler.read() to use 'wb' and 'rb' write/read modes respectively.  Windows makes a distinction between binary and text files so the 'b' mode extension is necessary.  Linux and other platforms ignore the 'b' and behave as expected so this update is cross platform.  When the data.py GeoElevationData.retrieve_or_load_file_data() method downloads the tile for the first time it returns the raw, unzipped data directly and everything works. But it also writes it in 'ascii' mode on Windows which causes errors when the data is read in subsequent calls to the file on disk.  This fix solves that issue.